feat!: add support for body-fitted elements
