+++ tetris/blockmesh/__init__.py
@@ -1,0 +1,16 @@
+# coding: utf-8
+"""Collection of python wrappers around various blockMesh elements."""
+
+from tetris.blockmesh.block import HexBlock
+from tetris.blockmesh.edge import (
+    ArcMidEdge,
+    ArcOriginEdge,
+    BSplineEdge,
+    LineEdge,
+    PolyLineEdge,
+    ProjectEdge,
+    SplineEdge,
+)
+from tetris.blockmesh.geometry import TriSurfaceMesh
+from tetris.blockmesh.patch import DefaultPatch, Face, Patch, PatchPair
+from tetris.blockmesh.vertex import ProjectVertex, Vertex
